Add files via upload
In this code we create an array and a pointer. Then point the pointer to the first[0 index ] element of the array. Then we take input from the user of the position of the array that has to be printed. Then we print the value at that position using pointer incrementation.